Transferring logic from a controller to a service

+++ src/main/java/com/bestSite/service/UserService.java
@@ -12,6 +12,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Collections;
@@ -21,6 +22,7 @@ import java.util.Optional;
 public class UserService {
 
     private User user;
+    private final Role roleAdmin = new Role(2L, "ADMIN");
     private final UserRepository userRepository;
     private final OverviewRepository overviewRepository;
     private final CloudService cloudService;
@@ -47,17 +49,25 @@ public class UserService {
 
     }
 
-    public void userUpdate(long id, User updatedUser, Optional<MultipartFile> newAvatar) {
+    public Model showUsers(Model model) {
+        Iterable<User> user = userRepository.findAll();
+        model.addAttribute("user", user);
+        return model;
+    }
+
+    public String userUpdate(long id, User updatedUser, BindingResult bindingResult, Optional<MultipartFile> newAvatar) {
+        if (bindingResult.hasErrors()) return "my-profile";
         user = userRepository.findById(id).orElseThrow();
         user.setFirstName(updatedUser.getFirstName());
         user.setLastName(updatedUser.getLastName());
         user.setEmail(updatedUser.getEmail());
-        if (cloudService.fileIsPresent(newAvatar.orElseThrow())){
+        if (cloudService.fileIsPresent(newAvatar.orElseThrow())) {
             String avatar = cloudService.uploadFile(newAvatar.orElseThrow());
             if (user.getAvatar() != null) cloudService.deleteFile(user.getAvatar());
             user.setAvatar(avatar);
         }
         userRepository.save(user);
+        return "redirect:/my-profile";
     }
 
     public Model showProfile(Model model, User user) {
@@ -65,6 +75,27 @@ public class UserService {
         model.addAttribute("user", user);
         model.addAttribute("overviews", overviews);
         return model;
+    }
+
+    public void makeOrRemoveAdmin(long id) {
+        user = userRepository.findById(id).orElseThrow();
+        if (user.getRoles().contains(roleAdmin)) {
+            user.removeRole(roleAdmin);
+        } else {
+            user.addRole(roleAdmin);
+        }
+        userRepository.save(user);
+    }
+
+    public void blockUnblock(long id) {
+        user = userRepository.findById(id).orElseThrow();
+        user.setStatus(user.getStatus().equals(Status.BLOCKED) ? Status.ACTIVE : Status.BLOCKED);
+        userRepository.save(user);
+    }
+
+    public void delete(long id) {
+        user = userRepository.findById(id).orElseThrow();
+        userRepository.delete(user);
     }
 
     public Authentication getCurrentUser() {
